requirements.txt
# Core Dependencies
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# APIs and HTTP
aiohttp==3.8.5

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2

# Scheduling
schedule==1.2.0
celery==5.3.1
redis==4.6.0

# Web Framework
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
alembic==1.11.3

# Monitoring
sentry-sdk==1.30.0
prometheus-client==0.17.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Additional dependencies found in code
jinja2==3.1.2  # For template rendering
pyyaml==6.0.1  # For YAML config files